refactor: extract bookings functions to utils
